+++ test/test_task_mnist.cpp
@@ -14,8 +14,8 @@ NANO_CASE(construction)
         const auto target_sum = scalar_t(2) - static_cast<scalar_t>(nano::size(odims));
 
         const auto folds = size_t(1);
-        const auto train_samples = 60000;
-        const auto test_samples = 10000;
+        const auto train_samples = size_t(60000);
+        const auto test_samples = size_t(10000);
         const auto train_fold = fold_t{0, protocol::train};
         const auto valid_fold = fold_t{0, protocol::valid};
         const auto test_fold = fold_t{0, protocol::test};
